requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
click>=8.1
requests>=2.28
beautifulsoup4>=4.11
lxml>=4.9
textstat>=0.7
scikit-learn>=1.3
nltk>=3.8
openai>=1.0
reportlab>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
